make command menu modular
