+++ components/command/index.tsx
@@ -1,47 +1,38 @@
 'use client';
 
-import { useState, useEffect, useCallback } from 'react';
-import { useRouter } from 'next/navigation';
 import { DialogTitle } from '@radix-ui/react-dialog';
 
+import { ChatHistoryGroup } from './group/chat-history';
+import { PagesGroup } from './group/navigation';
+import { WalletCommands } from './group/wallet';
+import { QuickActionsGroup } from './group/quick-actions';
+import { useCommandMenu } from './command-menu-context';
 import {
   CommandDialog,
   CommandEmpty,
-  CommandGroup,
   CommandInput,
   CommandItem,
   CommandList,
   CommandSeparator,
   CommandShortcut,
 } from '@/components/ui/command';
-import { ChatHistoryCommands } from './chat-history';
-import { commandMenuItems } from './items';
-import { WalletCommands } from './wallet';
+
+export type CommandMenuItem = {
+  icon: any;
+  label: string;
+  description: string;
+  url?: string;
+  disabled?: boolean;
+  keywords?: string[];
+};
+
+export type CommandMenuGroup = {
+  heading: string;
+  items: CommandMenuItem[];
+};
 
 export function CommandMenu() {
-  const [open, setOpen] = useState(false);
-  const router = useRouter();
-
-  useEffect(() => {
-    const down = (e: KeyboardEvent) => {
-      if (e.key === 'k' && (e.metaKey || e.ctrlKey)) {
-        e.preventDefault();
-        setOpen((open) => !open);
-      }
-    };
-    document.addEventListener('keydown', down);
-    return () => document.removeEventListener('keydown', down);
-  }, []);
-
-  const onSelect = useCallback(
-    (item: { url?: string }) => {
-      if (item.url) {
-        router.push(item.url);
-      }
-      setOpen(false);
-    },
-    [router],
-  );
+  const { open, setOpen } = useCommandMenu();
 
   return (
     <CommandDialog open={open} onOpenChange={setOpen}>
@@ -49,23 +40,9 @@ export function CommandMenu() {
       <CommandInput placeholder='Type a command or search...' />
       <CommandList>
         <CommandEmpty>No results found.</CommandEmpty>
-        {commandMenuItems.map((group, groupIndex) => (
-          <div key={group.heading}>
-            <CommandGroup heading={group.heading}>
-              {group.items.map((item) => (
-                <CommandItemComponent key={item.label} item={item} onSelect={() => onSelect(item)} />
-              ))}
-            </CommandGroup>
-            {groupIndex < commandMenuItems.length - 1 && <CommandSeparator />}
-          </div>
-        ))}
-        <CommandSeparator />
-        <ChatHistoryCommands
-          onSelect={(id) => {
-            router.push(`/chat/${id}`);
-            setOpen(false);
-          }}
-        />
+        <QuickActionsGroup />
+        <PagesGroup />
+        <ChatHistoryGroup />
         <CommandSeparator />
         <WalletCommands closeCommand={() => setOpen(false)} />
       </CommandList>
